Level Editor
John:
work on the level editor has begun, I have already added a bunch of
stuff, first of all the Level Editor extends Engine.

next i have set it up so that the user can look around the map with 
"WASD"

The user can also target a platfomr with "T and R" and move the platforn
arount with the arrow keys.