LevelEditor - hitbox Save and resize
John:
The level editor can now resize hitboxes and save them in the map file
so that when the game loads up it will load the correct hitbox.

Hitbox is now controlled via the "SHIFT" key, pressing shift allows you
to move the hitboxs via the arrow keys and to resize using "ZXCV" this
will likly change.