badguys - Damage Player
John:
badguys now create a damageHitbox when they attack. this is added to a
queue then the hitbox is compared to the player and if they overlap the
player takes damage. In the future I am going to add a player attack
queue as well or allow the badguys to hurt eachother.