Redundancy update.
John: now instead of storying a Sprite for every enemy. when a level is
loaded every sprite is added to a map with its name being a key, now
instead of haveing a buffered image for evey sprite on the map i have a
single sprite for every type of enemy. now when the game renders it
looks up the name of the target in the map and renders that.