Controller support canced
John:
I have decided not to focus on controllers support for release 0.1
for the V0.1 I want basic functionallity, for controller support I am
going to have to reseach a way to make a single multiplatform controller
driver that will be able to communicate with the main engine. If im
going to have controller support i want it on linux mac and PC. and i
dont want to have to release multiple versions. that goes against the
reason i choose java in the first place.