Sprites
Working on a new player sprite.

this might take a while ive already had to restart after about four
hours of work. but once this is finished i will be able to update the
player animations. after that i plan on fully implimeting sound then
running it through some stress tests and bugfixes. then the game should
be ready for lvl 0 v0.1 this well be used for finallizing how the game
is going to move and feel. 

I will probably open the game up to testers at this poiont to see how
smooth is runs and whatnot