Art time.
I have reached a point where it is time to begin focusing on artwork and
level design, because of this I will be spendin my time making the game
look and play good. so it might look like I have given up, but i havent.

due to the nature of the internet I will not be uploading the sprite
that will be in the final version to github. instead i will be backing
them up through other means.