Engine - Input Update
John:
I've started the bigginning of the inputs so for the new combat system.
basically it will bass a String of letters, and if a letter is there
that a true of its not then false,so if there is an 'N' then the player
is pressing the up key, 'L' light attack.

v 1.0.189