Trigger engine fix
John:
just a quick fix for the trigger engine. there was a couple lines that
were causing issues with the JAR version of the game. its all fixed now
so time to start working on art again :)
this is going to be my last commit for a while now while I worry about
artwork and level design :). but dont worry ill be back to work on the
soldiers and bosses :)

V 1.1.301